refactor(test): extract shared guardrail helpers to non-test modules (#1431)
Two tests/unit/ behavioral files imported helpers from tests/_guardrails/
*test* modules — a test-imports-test smell. Extract the shared, reusable
helpers into two new ``_``-prefixed NON-test modules that both the gate
files and the behavioral files import from:

- tests/_guardrails/_ast_reach_in.py: the AST reach-in visitors / accessor
  helpers (``_RuntimeImportVisitor`` + ``_facade_construction_lines`` and the
  ``_self_attr_*`` / ``_method_body`` / ``_call_keyword_value`` accessors plus
  their private transitive helpers) shared with test_init_order.py.
- tests/_guardrails/_cassette_shape_lint.py: the leak-pattern detectors and
  ``DISPLAY_NAME_FALSE_POSITIVES`` allowlist (with ``_find_leaks``) shared with
  test_cassette_patterns.py.

test_no_facade_reach_in.py and test_cassette_shapes.py now import their helpers
from the new modules; test_init_order.py and test_cassette_patterns.py import
from the new modules too, so neither behavioral test imports from another test
module. The moved code is byte-identical and all assertions are unchanged.

``_``-prefixed files are not collected by pytest (their names don't match the
``test_*.py`` glob), so the helpers never run as tests.
